Replace UCHAR, UINT, USHORT, and ULONG

+++ include/rtw_mp.h
@@ -164,24 +164,24 @@ typedef struct _MPT_CONTEXT
 	MPT_WORK_ITEM_HANDLER	CurrMptAct;
 
 	// 1=Start, 0=Stop from UI.
-	ULONG			MptTestStart;
+	u32			MptTestStart;
 	// _TEST_MODE, defined in MPT_Req2.h
-	ULONG			MptTestItem;
+	u32			MptTestItem;
 	// Variable needed in each implementation of CurrMptAct.
-	ULONG			MptActType;	// Type of action performed in CurrMptAct.
+	u32			MptActType;	// Type of action performed in CurrMptAct.
 	// The Offset of IO operation is depend of MptActType.
-	ULONG			MptIoOffset;
+	u32			MptIoOffset;
 	// The Value of IO operation is depend of MptActType.
-	ULONG			MptIoValue;
+	u32			MptIoValue;
 	// The RfPath of IO operation is depend of MptActType.
-	ULONG			MptRfPath;
+	u32			MptRfPath;
 
 	WIRELESS_MODE		MptWirelessModeToSw;	// Wireless mode to switch.
 	u8			MptChannelToSw;		// Channel to switch.
 	u8			MptInitGainToSet;	// Initial gain to set.
-	//ULONG			bMptAntennaA;		// TRUE if we want to use antenna A.
-	ULONG			MptBandWidth;		// bandwidth to switch.
-	ULONG			MptRateIndex;		// rate index.
+	//u32			bMptAntennaA;		// TRUE if we want to use antenna A.
+	u32			MptBandWidth;		// bandwidth to switch.
+	u32			MptRateIndex;		// rate index.
 	// Register value kept for Single Carrier Tx test.
 	u8			btMpCckTxPower;
 	// Register value kept for Single Carrier Tx test.
@@ -190,13 +190,13 @@ typedef struct _MPT_CONTEXT
 	u8			TxPwrLevel[2];	// rf-A, rf-B
 
 	// Content of RCR Regsiter for Mass Production Test.
-	ULONG			MptRCR;
+	u32			MptRCR;
 	// TRUE if we only receive packets with specific pattern.
 	bool			bMptFilterPattern;
 	// Rx OK count, statistics used in Mass Production Test.
-	ULONG			MptRxOkCnt;
+	u32			MptRxOkCnt;
 	// Rx CRC32 error count, statistics used in Mass Production Test.
-	ULONG			MptRxCrcErrCnt;
+	u32			MptRxCrcErrCnt;
 
 	bool			bCckContTx;	// TRUE if we are in CCK Continuous Tx test.
 	bool			bOfdmContTx;	// TRUE if we are in OFDM Continuous Tx test.
@@ -210,7 +210,7 @@ typedef struct _MPT_CONTEXT
 
 	// ACK counter asked by K.Y..
 	bool			bMptEnableAckCounter;
-	ULONG			MptAckCounter;
+	u32			MptAckCounter;
 
 	// SD3 Willis For 8192S to save 1T/2T RF table for ACUT	Only fro ACUT delete later ~~~!
 	//s1Byte		BufOfLines[2][MAX_LINES_HWCONFIG_TXT][MAX_BYTES_LINE_HWCONFIG_TXT];
